feat: Sync GraphQL schema for decimal quantity support (issue #132)
Phase 2: GraphQL Schema Sync + Codegen

Backend changes:
- BuildMutation.cs: AddPart mutation accepts decimal quantity
- ValidationService.cs: ValidateQuantity accepts decimal
- schema.graphql: Added Decimal scalar + updated Part/mutation types

Frontend changes:
- codegen.ts: Configure Decimal scalar → number mapping
- build.graphql: AddPart mutation uses Decimal input type
- graphql.ts: Regenerated with Decimal type support

Schema updates:
- scalar Decimal definition
- Part.quantity: Int! → Decimal!
- addPart($quantity: Decimal!): Part!

Type safety: Full end-to-end decimal support (C# → schema → graphql.ts)

Build verification:
✅ Backend: dotnet build succeeds
✅ Frontend: pnpm codegen succeeds
✅ Type sync: Decimal throughout pipeline

Next: Phase 3 - Test coverage + edge case tests

+++ backend/src/FactoryApp.GraphQL/Services/ValidationService.cs
@@ -58,7 +58,7 @@ public static class ValidationService
             throw new GraphQLException("SKU must be <= 100 characters");
     }
 
-    public static void ValidateQuantity(int quantity)
+    public static void ValidateQuantity(decimal quantity)
     {
         if (quantity <= 0)
             throw new GraphQLException("Quantity must be greater than 0");
